- fix menu re filling process (multisite)
git-svn-id: http://svn.civicrm.org/civicrm/trunk@22930 cff41f8d-9fe8-0310-9d4e-fb156a3ffe5e

+++ CRM/Core/Menu.php
@@ -233,11 +233,12 @@ class CRM_Core_Menu
         self::buildAdminLinks( $menu );
     }
 
-    static function store( ) {
+    static function store( $truncate = true ) {
         // first clean up the db
-        $query = 'TRUNCATE civicrm_menu';
-        CRM_Core_DAO::executeQuery( $query );
-
+        if ( $truncate ) {
+            $query = 'TRUNCATE civicrm_menu';
+            CRM_Core_DAO::executeQuery( $query );
+        }
         $menu =& self::items( );
 
         self::build( $menu );
